fix rename SidebarContext file
