+++ src/components/NavBar/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import { useSideBarContext } from "../../context/SideBarComponentContext";
+import { useSideBarContext } from "../../context/SidebarContext";
 import Icon from "../Icon";
 import { IconNames } from "../Icon/IconNames";
