+++ src/components/styledComponents/MainContainer.tsx
@@ -1,6 +1,6 @@
 import { PropsWithChildren } from "react";
 import styled from "styled-components";
-import { useSideBarContext } from "../../context/SideBarComponentContext";
+import { useSideBarContext } from "../../context/SidebarContext";
 
 export const Container = styled.div`
   margin-left: ${({ expanded }: { expanded?: boolean }) =>
